Use product flavours to enable builds without proprietary libraries
This commit uses build flavours to provide two different builds: A
'floss' build and a 'google' build.

- The 'floss' build builds a basic version of the app with no
  dependencies on proprietary libraries.
- The 'google' build adds Chromecast support and a few other extras
  provided by Google Play Services
