+++ app/src/main/AndroidManifest.xml
@@ -243,18 +243,7 @@
 
         <meta-data android:name="android.app.default_searchable"
                    android:value="github.daneren2005.dsub.activity.QueryReceiverActivity"/>
-                   
-        <meta-data android:name="com.google.android.backup.api_key"
-        	android:value="AEdPqrEAAAAIUhOMtwa_eG-f0oYUHnetl_Cz7cO9zae8ZXOK5w"/>
 
-		<meta-data
-			android:name="com.google.android.gms.version"
-			android:value="@integer/google_play_services_version" />
-
-		<meta-data android:name="com.google.android.gms.car.application"
-				   android:resource="@xml/auto_app_description"/>
-		<meta-data android:name="com.google.android.gms.car.notification.SmallIcon"
-				   android:resource="@drawable/stat_notify_playing" />
 	</application>
 
 </manifest>
